refactor: 파이어베이스 연결 공통 함수 구현
