+++ lib/enum/service_name.dart
@@ -1,0 +1,16 @@
+enum ServiceName {
+  /* movie 전체 데이터 */
+  MOVIE,
+
+  /* movie 안에 content 데이터 */
+  CONTENT,
+
+  /* movie 안에 like 데이터 */
+  LIKE,
+
+  /* movie 안에 poster 데이터 */
+  POSTER,
+
+  /* movie 안에 title 데이터 */
+  TITLE,
+}
