+++ lib/screen/like_screen.dart
@@ -1,3 +1,5 @@
+import 'package:clone_netflix/enum/service_name.dart';
+import 'package:clone_netflix/firebase/data_store.dart';
 import 'package:clone_netflix/layout/base_layout.dart';
 import 'package:clone_netflix/model/movieData_model.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -13,42 +15,42 @@ class LikeScreen extends StatefulWidget {
 }
 
 class _LikeState extends State<LikeScreen> {
-  Widget rebuildBody(BuildContext context) {
-    return StreamBuilder<QuerySnapshot>(
-      stream: Firestore.instance
-          .collection('movie')
-          .where('like', isEqualTo: true)
-          .snapshots(),
-      builder: (context, snapshot) {
-        if (!snapshot.hasData) return LinearProgressIndicator();
-        return buildList(context, snapshot.data!.documents);
-      },
-    );
-  }
-
-  Widget buildList(context, snapshot) {
-    return Expanded(
-      child: GridView.count(
-        crossAxisCount: 3,
-        childAspectRatio: 1 / 1.5,
-        padding: EdgeInsets.all(3.0),
-        children: snapshot.map((data) => posterItem(context, data)).toList(),
-      ),
-    );
-  }
-
-  Widget posterItem(context, data) {
-    final movieData = MovieData.fromSnapshot(data);
-    return InkWell(
-      child: Padding(
-        padding: EdgeInsets.only(right: 8.0),
-        child: Image.asset(movieData.poster),
-      ),
-      onTap: () {
-        Get.toNamed('/netflixDetail', arguments: movieData);
-      },
-    );
-  }
+  // Widget rebuildBody(BuildContext context) {
+  //   return StreamBuilder<QuerySnapshot>(
+  //     stream: Firestore.instance
+  //         .collection('movie')
+  //         .where('like', isEqualTo: true)
+  //         .snapshots(),
+  //     builder: (context, snapshot) {
+  //       if (!snapshot.hasData) return LinearProgressIndicator();
+  //       return buildList(context, snapshot.data!.documents);
+  //     },
+  //   );
+  // }
+  //
+  // Widget buildList(context, snapshot) {
+  //   return Expanded(
+  //     child: GridView.count(
+  //       crossAxisCount: 3,
+  //       childAspectRatio: 1 / 1.5,
+  //       padding: EdgeInsets.all(3.0),
+  //       children: snapshot.map((data) => posterItem(context, data)).toList(),
+  //     ),
+  //   );
+  // }
+  //
+  // Widget posterItem(context, data) {
+  //   final movieData = MovieData.fromSnapshot(data);
+  //   return InkWell(
+  //     child: Padding(
+  //       padding: EdgeInsets.only(right: 8.0),
+  //       child: Image.asset(movieData.poster),
+  //     ),
+  //     onTap: () {
+  //       Get.toNamed('/netflixDetail', arguments: movieData);
+  //     },
+  //   );
+  // }
 
   @override
   Widget build(BuildContext context) {
@@ -76,7 +78,7 @@ class _LikeState extends State<LikeScreen> {
                 ],
               ),
             ),
-            rebuildBody(context),
+            DataStore(serviceName: ServiceName.LIKE),
           ],
         ),
       ),
